review: rename remote_prover to proving_service, rename request and response structs from proto
